backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
Django==5.2.11
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-auth==2.48.0
google-auth-oauthlib==1.3.0
gspread==6.2.1
idna==3.11
oauthlib==3.3.1
pillow==12.1.1
psycopg==3.3.3
psycopg-binary==3.3.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.11.0
python-decouple==3.8
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
